Adding bar range markers (including signaling system)
Adding bar range markers (like to denote what the range of the bar chart
axis is), which required having a system to add padding to the beginning
of tracks (to make space for the markers), which is an oncoprint-level
change that must be initiated by a cell renderer (maybe should change to
initiate by a track renderer...). So there's a signal system where
signals get passed up to the oncoprint and then events get trickled down.